[Fix] Rename object_delete_f callback to object_destructor_f.
The rename is to make the semantics more clear. The object_new_f
callback is the real 'new' operation: allocation + initialization, while
object deleting is different because the lifetime management logic is
fully delegated to esch_object_delete(). When an object is being
deleted, esch_object_delete() determines when to do real deleting, while
esch_alloc determines how to deallocate the memory. The only step
required from customized object, is to provide a destructor to clean up
internal logic.